[AMD] Rename aiter_unified_attention_mtp to unified_attention_3d_mtp
The module holds a Qwen3.5 MTP-verify specialization of the Triton
kernel_unified_attention_3d launch; name it after the kernel it
specializes rather than after AITER.
